oracle-brain/requirements.txt
anthropic
httpx
websockets
python-dotenv
fastapi
uvicorn[standard]
